+++ Utilities/gdcm/Source/MediaStorageAndFileFormat/gdcmLookupTable.h
@@ -50,7 +50,7 @@ public:
     unsigned short subscript, unsigned short bitsize);
   unsigned int GetLUTLength(LookupTableType type) const;
   virtual void SetLUT(LookupTableType type, const unsigned char *array,
-    unsigned int length);
+    size_t length);
   void GetLUT(LookupTableType type, unsigned char *array, unsigned int &length) const;
   void GetLUTDescriptor(LookupTableType type, unsigned short &length,
     unsigned short &subscript, unsigned short &bitsize) const;
